WIP: Everything is in a decent state
Vertical Collection itself is working. Next steps will be to refactor the
KeyList and Radar concepts to make them more self contained, and remove
as much of the tracking logic from teh component itself as possible. Also
need to rethink some of the recycling logic, it doesn't work if we want to
have buffers and to not remove components. Also, we should rethink silentNight
it'll be necessary for both prepending and changing heights from above.