refactor(number-tree): simplifies the architecture
This refactor rethinks the way that we track the positions of elements
that are not rendered. Rather tracking absolute values and updating them
whenever we scroll, we track positional values in a BST/Skip List-like
data structure. Given an offset from the top of the scrollable area, the
structure can return the nearest visible index, the heightBefore, and the
heightAfter the index.

From there everything is relational math. We don't need the construct of the
Radar any more for scroll handling, as the only thing we need to track is our
scrollable area's offset relative to the document (if the document is the
container).