refactor(radar): Converts `radar` into a computed property
It turns out that `radar` needs to be a computed property in order to respond to
array mutations. Something in the template itself has to be bound in order for radar
to be recomputed, and as the only thing that is bound at all is `virtualComponents`,
they needed to be moved into `radar` proper to put it on the refresh path.