feature(resize): Adds a resize event
Even when we are always remeasuring, there is an edge case that we can't easily account for.
If a large number of items are rendered, and then all of them are shrunk (or even one very
large item is shrunk), it's possible that A. no scroll event is fired (sometimes it happens
when we resize items, sometimes it doesn't) and B. the buffers aren't large enough to account
for the delta, and we end up showing the padding below.

This adds a "resize" event based on the concept found here:

https://github.com/marcj/css-element-queries/blob/master/src/ResizeSensor.js

The basic idea is you listen to the scroll event on a hidden container that is a child of the
element you are trying to measure, with `width: 100%` and `height: 100%`. It has a child with
a very large height/width, and when it resizes the scroll bars change size triggering the
scroll event.

This is pretty hack-y, but it seems to be the best solution out there.