bugfix(scheduling): Moves all scheduling into the RAF and ensures proper cancelling
Because of the setTimeout bug described in #31, we cannot reliably schedule using setTimeout/
run.next(). There are two places where this happens - 1st in the component itself when it
schedules sending actions, and second in the DynamicRadar when it schedules a full measure.
The full measure appears to be redundant, as another full measure will be done before any changes
occur in the next scroll (per the change in measuring strategy to accomodate item height changes)
and the action sending can be done in the RAF by scheduling on the curren runloop (Todo: Add a
perf notes/gotchas page in the docs to go over the implications of this design choice).

Also plugs some cancellation bugs that were probably causing the failures in #33.

Closes #31 and #33