feat(scaling): Makes the collection work with scaling
When used in a container with `transform: scale`, the collection would
break due to measurements being off. For example, we would measure an
item that was normally 20px to be scaled by say 10px, and then we would
render the measured value, which would be scaled again down to 5px. Lots
of small bugs popped up due to this.

This PR adds a `getScaledClientRect` helper which accepts a scale value
and element and returns the client rect scaled to the correct size. We
determine the scale value in `updateConstants` by comparing the client
rect of the scroll container to it's offsetHeight.

This strategy does not account for cases where, for instance, the item
container is scaled, but the scroll container isn't. We can probably
guard against these cases in debug assertions.