feat(tests): Adds a test waiter for the RAF scheduler
Adds a test waiter that polls until the scheduler's job queue is empty.

The reason for using the `registerWaiter` API despite it's warts is
ultimately that it really is the only way to conveniently integrate with
the `wait` helper. This has a few implications:

* It's the easiest way to support legacy test helpers (globals)
* It's the only way to completely prevent the user from having to think
about whether an action may-or-may-not schedule a RAF. Ex:

```js
test('test a scrolling thing', async function() {
  scrollTo('.scroll-container', 0, 100);

  // ok, that definitely schedules a RAF, better wait
  await waitForRender();

  click('.button-that-causes-scroll')

  // that'll probably cause a RAF, better wait
  await waitForRender();

  fillIn('.search', 'query');

  // that... might cause a RAF? Better wait
  await waitForRender();
});
```

We end up wanting to wait for most things anyways, or at the least not
having to worry about waiting.