refactor(constants): Refactors estimateElementHeight and fixes minor bugs
This refactor addresses several concerns that were leaking into the
transform-scale PR:

1. Refactors `estimateElementHeight` and `estimateElementMaxHeight` to
be a single function with more well defined behavior. Also makes
`estimateElementHeight` use the `offsetHeight` of the parent element
in the percent case, meaning we won't need to have the transform scale
value at all.

2. Enforces that `estimateHeight` must evaluate to some value > 0,
rather than the previous behavior where it would simply default to `1`.

3. Passes the correct elements into `estimateElementHeight` for
estimating the scrollContainer's max-height, it should be basing the
estimate off the scrollContainer's parent

4. Adds tests for the max-height behavior with various units

5. renames `_estimateHeight` and `_scrollContainerHeight` to
`_calculatedEstimateHeight` and `_calculatedScrollContainerHeight` to
more clearly reflect what the values actually are