feat(dynamic-pool): Implements a dynamic component pool with incremental render (#151)
* feat(dynamic-pool): Implements a dynamic component pool with incremental render

This PR returns vertical-collection to a dynamic component pool, adding
and removing components from the DOM as necessary based on actual
component measurements. The goal is that we never render more items than
we need to, allowing the collection to respond to varying sizes of items
(e.g. items that are much larger or smaller than the estimate height).

The last time this strategy was attempted it was reverted due to being
a performance regression. Destroying and re-creating components
dynamically cost too much compared to the benefits. This time we're
getting around that issue by never actually destroying any components.
We simply remove their DOM, appending it to a temporary container
element and then re-appending it when we need another component.

An additional feature required to make this fully work is incremental
render. If enabled, the collection will measure and schedule another
render if it turned out it was wrong about the number of items needed,
for instance because they were too small.

* measure -> sync

* guard against calling incremental render twice

* null out next incremental render on cancel