bugfix(timing): Pulls prepend/append logic into Radar (#168)
Pulls the prepend/append/reset logic into Radar due to issues with
timing guarantees. This guaratees that we will check the state of the
array when we are going to update, not when we are scheduling the
change. In certain edge cases, updates to the array can occur between
when the computed property is calculated and the Radar actually updates.