[bugfix] Fix fastboot and refactor occluded-content (#193)
This PR fixes fastboot and refactors element proxies to separate
concerns more clearly. OccludedContent is now a separate class with
different behavior than VirtualComponent (which should also be updated
in the future, needs better naming) and more direct proxy-ing to its
underlying element. Proxying like this allows us to check for the
element's existence before we continue with any operations, meaning
we're fastboot-safe.

Not every usage of `document` has been wrapped in fastboot checks, but
all of the remaining ones do not occur in fastboot anyways - updates are
first scheduled on `didInsertElement` so we're safe for the most part
here.