Use once instead of setTimeout + run
In our test environment we were getting an error:
`Cannot call get with 'id' on an undefined object`.
It was quite flaky, happening only occasionally.
However we've noticed that sometimes it is happening in production as well.
There is an open issue in the repo for that bug:
https://github.com/html-next/vertical-collection/issues/317

Since this is a flaky issue it is hard to write a test for it, however
there was one test in our test suite which was flaking more frequently,
so, we took it as reference point to check if the fix is working.
But the test is not easy to isolate and integrate into
vertical-collection repo, so, this commit won't include test as a proof
for the fix.

====

The actual change:

This bug was happening when vertical collection component was trying
to send actions to the users. It was being scheduled through setTimeout
and wrapped in run(). setTimeout() is not bound to the ember runloop,
this is why explicit `run` was needed.

My guess is that because of that the send actions code could be run by
the moment when the item for which the action was meant to be sent
had gone from the vertical collection.

In fact Ember has API to schedule some code to be run within the proper
queue which is the `actions` queue in this particular case. Using that
API seems to be fixing `Cannot call get with 'id' on an undefined object`
exception for us.