Don't send action on non-existing items
In our test environment we were getting an error:
`Cannot call get with 'id' on an undefined object`.
It was quite flaky, happening only occasionally.
However we've noticed that sometimes it is happening in production as well.
There is an open issue in the repo for that bug:
https://github.com/html-next/vertical-collection/issues/317

Check that the item the event is being generated for still exists.

Here’s a link to the discussion of the same issue in a separate PR:
https://github.com/html-next/vertical-collection/pull/324#issuecomment-755324980