delete virtual domPool element
The radar has a domPool virtual element to save temporarily deleted nodes and reuse them when a new item is added. The domPool is used only if the component has fewer items than possible rendered visible items after the update. So if the item is deleted from the items list the item's node can still exist in the virtual dom so the item's component can be still updated.

This behavior isn't working for computed fields on the ember models. Ember rejects the error when trying to access the computed property on the destroyed object if it hasn't been consumed before. This is possible if the items' component should render more fields after the update.

As this behavior works only on a small items list it doesn't affect scroll rendering, so it gives a little impact on performance but gives unexpected errors while updating deleted items so we can get rid of this and delete unused nodes if they can't be reused right now.