clear scheduled actions after recalculation
Scheduled actions as the name says are fired with delay after the render is finished. They are going through the `sync` queue (`scheduleUpdate` method) and the `measure` queue (`update` method) in the radar class and `setTimeout` (`_scheduleSendAction` method) in the component. So there is some gap before the event is actually fired when the `items` argument can be changed and the event will be fired with an exception. The scheduled actions include the event and index of the item, but due to timeout, the actual item may not exist for this index.
Clear the timeout and the queue before scheduling a new update after recalculation to omit these outdated events and fire only the new actual events